+++ R/namespace.R
@@ -330,7 +330,7 @@ roxy_tag_ns.roxy_tag_import <- function(x, block, env) {
 
 #' @export
 roxy_tag_parse.roxy_tag_importClassesFrom <- function(x) {
-  tag_words(x, min = 2)
+  tag_words(x, min = 2, multiline = "indent")
 }
 #' @export
 roxy_tag_ns.roxy_tag_importClassesFrom <- function(x, block, env) {
@@ -339,7 +339,7 @@ roxy_tag_ns.roxy_tag_importClassesFrom <- function(x, block, env) {
 
 #' @export
 roxy_tag_parse.roxy_tag_importFrom <- function(x) {
-  tag_words(x, min = 2)
+  tag_words(x, min = 2, multiline = "indent")
 }
 #' @export
 roxy_tag_ns.roxy_tag_importFrom <- function(x, block, env) {
@@ -369,7 +369,7 @@ roxy_tag_ns.roxy_tag_importFrom <- function(x, block, env) {
 
 #' @export
 roxy_tag_parse.roxy_tag_importMethodsFrom <- function(x) {
-  tag_words(x, min = 2)
+  tag_words(x, min = 2, multiline = "indent")
 }
 #' @export
 roxy_tag_ns.roxy_tag_importMethodsFrom <- function(x, block, env) {